fixing a bug

+++ src/cpp/dxvalidator/dxvalidate_contigset.cc
@@ -269,12 +269,12 @@ int ContigSetReader::chrIndex(const string &name) {
   return it->second;
 }
 
-int ContigSetReader::chrSize(int i) {
+int64_t ContigSetReader::chrSize(int i) {
   if ((i < 0) || (i >= sizes.size())) return -1;
   return sizes[i];
 }
 
-int ContigSetReader::chrOffset(int i) {
+int64_t ContigSetReader::chrOffset(int i) {
   if (! hasOffset) return -1;
   if ((i < 0) || ( i >=offsets.size())) return -1;
   return offsets[i];
